Extract PDF import spreadsheet helpers
