+++ tests/test-audit.js
@@ -73,7 +73,8 @@ assert('SW APP_SHELL includes PDF import review module', swAuditSrc.includes("'/
 assert('SW APP_SHELL includes modal lifecycle module', swAuditSrc.includes("'/js/modal-lifecycle.js'"));
 assert('SW APP_SHELL includes marker analysis module', swAuditSrc.includes("'/js/marker-analysis.js'"));
 assert('SW APP_SHELL includes PDF import support modules',
-  swAuditSrc.includes("'/js/pdf-import-preflight.js'")
+  swAuditSrc.includes("'/js/pdf-import-spreadsheet.js'")
+  && swAuditSrc.includes("'/js/pdf-import-preflight.js'")
   && swAuditSrc.includes("'/js/pdf-import-progress.js'")
   && swAuditSrc.includes("'/js/pdf-import-marker-normalization.js'")
   && swAuditSrc.includes("'/js/pdf-import-persistence.js'"));
